// web/src/utils/enum.ts
export enum VisibilityType {
    None = 0,
    Modal = 1,
    VehicleStorage = 2,
    TruckerPDA = 3,
    TruckRentalMenu = 4
}

export enum PDAVisibilityType {
    Statics = 0,
    IndustryList = 1,
    VehicleCapacityList = 2
}

export enum IndustryTier {
    Primary = 1,
    Secondary = 2,
    Business = 3
}

export enum ModalType {
    Confirm = 0,
    Dialog = 1
}